REFACTOR: extract `UseHide` bc the fqname is toooo long
